[Refactor] RectangleView -> BasicShapeView를 부모로 갖는 별도의 클래스 타입으로 변경
